Fix iTunes upload error:
Unexpected CFBundleExecutable Key - The bundle at '/Payload/mymobilenative.app/app/tns_modules/nativescript-local-notifications/native-src/ios/LocalNotificationsPlugin/Info.plist' does not contain a bundle executable. If this bundle intentionally does not contain an executable, consider removing the CFBundleExecutable key from its Info.plist and using a CFBundlePackageType of BNDL. If this bundle is part of a third-party framework, consider contacting the developer of the framework for an update to address this issue.